requirements.txt
asgiref==3.11.0
Django==5.2.8
django-simple-history==3.10.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
